+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/cards.js
@@ -224,6 +224,7 @@ if (!gwoCardsLoaded) {
           ];
           var expandedCards = [
             "gwaio_enable_planetaryradar",
+            "gwaio_speed_structure",
             "gwaio_upgrade_advancedairfactory",
             "gwaio_upgrade_advancedbotfactory",
             "gwaio_upgrade_advancedenergyplant",
@@ -465,8 +466,6 @@ if (!gwoCardsLoaded) {
                     if (match) {
                       cardChance.chance = 0;
                     }
-
-                    console.log(card.id, cardChance);
 
                     return cardChance;
                   });